+++ src/constants.ts
@@ -4,12 +4,14 @@ import {
   FRANCE_COUNTRY,
   GERMANY_COUNTRY,
   SPAIN_COUNTRY,
+  SWITZERLAND_COUNTRY,
 } from "./geo/countries.js";
 
 export const SPAIN_CENTER: [number, number] = SPAIN_COUNTRY.center;
 export const GERMANY_CENTER: [number, number] = GERMANY_COUNTRY.center;
 export const FRANCE_CENTER: [number, number] = FRANCE_COUNTRY.center;
 export const DENMARK_CENTER: [number, number] = DENMARK_COUNTRY.center;
+export const SWITZERLAND_CENTER: [number, number] = SWITZERLAND_COUNTRY.center;
 export const CZECHIA_CENTER: [number, number] = COUNTRIES.CZ.center;
 export const POLAND_CENTER: [number, number] = COUNTRIES.PL.center;
 export const DEFAULT_ZOOM = 6;